requirements.txt
# OrganoidReader Requirements
# Core dependencies for AI-powered organoids image analysis

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0
imageio>=2.31.0
albumentations>=1.3.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# GUI Framework
PyQt5>=5.15.0
# Alternative: PyQt6>=6.5.0

# Data Visualization
plotly>=5.15.0
bokeh>=3.2.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Image Format Support
tifffile>=2023.7.0
czifile>=2019.7.2
nd2reader>=3.3.0

# Configuration and Logging
pyyaml>=6.0
configparser>=6.0.0
loguru>=0.7.0

# Database
# sqlite3  # Built-in Python module no need installation

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Performance Monitoring
memory-profiler>=0.61.0
line-profiler>=4.1.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=1.0.0